refactor graph.jsx; separate cy.js api usage + extensions from app logic - initial refactor towards integrating the viewer - adds button to choose a layout - adds layout state that the button alters on click
